refactor(db): change package name
- Updated the database connection logic to utilize a new custom `db` package instead of gorm.
- Refactored the API and mail server to establish database connections using the new package.
- Modified database queries to use raw SQL statements for better control and performance.
